pyproject.toml
[tool.poetry]
name = "cs224g"
version = "0.1.0"
description = ""
authors = [
    "winsonc7 <winsonc@stanford.edu>"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10"
openai = ">=1.61.1,<2.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
flask = ">=3.1.0,<4.0.0"
flask-cors = ">=4.0.0,<5.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
